third_party/WebKit/LayoutTests/media/video-autoplay-experiment-modes-expected.txt

CONSOLE WARNING: Failed to execute 'play' on 'HTMLMediaElement': API can only be initiated by a user gesture.
CONSOLE ERROR: line 71: Uncaught (in promise) NotAllowedError: play() can only be initiated by a user gesture.
CONSOLE WARNING: Failed to execute 'play' on 'HTMLMediaElement': API can only be initiated by a user gesture.
CONSOLE ERROR: line 71: Uncaught (in promise) NotAllowedError: play() can only be initiated by a user gesture.
CONSOLE WARNING: Failed to execute 'play' on 'HTMLMediaElement': API can only be initiated by a user gesture.
CONSOLE ERROR: line 71: Uncaught (in promise) NotAllowedError: play() can only be initiated by a user gesture.
CONSOLE WARNING: Failed to execute 'play' on 'HTMLMediaElement': API can only be initiated by a user gesture.
CONSOLE ERROR: line 71: Uncaught (in promise) NotAllowedError: play() can only be initiated by a user gesture.
CONSOLE WARNING: Failed to execute 'play' on 'HTMLMediaElement': API can only be initiated by a user gesture.
CONSOLE ERROR: line 71: Uncaught (in promise) NotAllowedError: play() can only be initiated by a user gesture.
CONSOLE WARNING: Failed to execute 'play' on 'HTMLMediaElement': API can only be initiated by a user gesture.
CONSOLE ERROR: line 71: Uncaught (in promise) NotAllowedError: play() can only be initiated by a user gesture.
CONSOLE WARNING: Failed to execute 'play' on 'HTMLMediaElement': API can only be initiated by a user gesture.
CONSOLE ERROR: line 71: Uncaught (in promise) NotAllowedError: play() can only be initiated by a user gesture.
CONSOLE WARNING: Failed to execute 'play' on 'HTMLMediaElement': API can only be initiated by a user gesture.
CONSOLE ERROR: line 71: Uncaught (in promise) NotAllowedError: play() can only be initiated by a user gesture.
CONSOLE WARNING: Failed to execute 'play' on 'HTMLMediaElement': API can only be initiated by a user gesture.
CONSOLE ERROR: line 71: Uncaught (in promise) NotAllowedError: play() can only be initiated by a user gesture.
END OF TEST
  Check if the autoplay gesture override experiment works.  There are a lot
  of config options, so this test just runs all of them.

  The "results" table contains one row per config tested.
  == Test Inputs ==
  #      - config number, in case you'd like to run just one.
  Flags  - autoplay experiment setting being tested.
               a      - "foraudio"
               v      - "forvideo"
               V      - "ifviewport"
               P      - "ifpagevisible"
               M      - "ifmuted"
               p      - "playmuted"
               m      - "ifmobile"
               For example, vM means '-forvideo-ifmuted".
  Type   - audio or video element?
               audio  - 
#	Flags	Type	Play w/	Mute	Mobile	View	Setting	Early?	Played?	Muted?
0		video	none	no	no	onscree	enabled	-	no	-
1	v	video	none	no	no	onscree	enabled	-	no	-
3	vVM	video	none	no	no	onscree	enabled	-	no	-
4	vVp	video	none	no	no	onscree	enabled	-	no	-
7		video	attr	no	no	onscree	enabled	-	no	-
8	v	video	attr	no	no	onscree	enabled	-	played	unmuted
9	vP	video	attr	no	no	onscree	enabled	-	played	unmuted
10	vVM	video	attr	no	no	onscree	enabled	-	no	-
11	vVp	video	attr	no	no	onscree	enabled	-	played	muted
12	vm	video	attr	no	no	onscree	enabled	-	no	-
13	a	video	attr	no	no	onscree	enabled	-	no	-
14		video	play()	no	no	onscree	enabled	-	no	-
15	v	video	play()	no	no	onscree	enabled	-	played	unmuted
17	vVM	video	play()	no	no	onscree	enabled	-	no	-
18	vVp	video	play()	no	no	onscree	enabled	-	played	muted
24	vVM	video	none	yes	no	onscree	enabled	-	no	-
31	vVM	video	attr	yes	no	onscree	enabled	-	played	muted
38	vVM	video	play()	yes	no	onscree	enabled	-	played	muted
49		video	attr	no	no	scroll	enabled	no	no	-
50	v	video	attr	no	no	scroll	enabled	yes	played	unmuted
53	vVp	video	attr	no	no	scroll	enabled	no	played	muted
56		video	play()	no	no	scroll	enabled	no	no	-
57	v	video	play()	no	no	scroll	enabled	yes	played	unmuted
60	vVp	video	play()	no	no	scroll	enabled	no	played	muted
91		video	attr	no	no	offscre	enabled	-	no	-
92	v	video	attr	no	no	offscre	enabled	-	played	unmuted
95	vVp	video	attr	no	no	offscre	enabled	-	no	-
135	vP	video	attr	no	no	obscure	enabled	-	no	-
182		video	play()	no	no	onscree	disable	-	no	-
183	v	video	play()	no	no	onscree	disable	-	no	-
185	vVM	video	play()	no	no	onscree	disable	-	no	-
186	vVp	video	play()	no	no	onscree	disable	-	no	-
206	vVM	video	play()	yes	no	onscree	disable	-	no	-
344	v	video	attr	no	yes	onscree	enabled	-	played	unmuted
348	vm	video	attr	no	yes	onscree	enabled	-	played	unmuted
